+++ packages/runtime-core/__tests__/componentEmits.spec.ts
@@ -143,12 +143,78 @@ describe('component: emit', () => {
     expect(`event validation failed for event "foo"`).toHaveBeenWarned()
   })
 
-  test('isEmitListener', () => {
-    expect(isEmitListener(['click'], 'onClick')).toBe(true)
-    expect(isEmitListener(['click'], 'onclick')).toBe(false)
-    expect(isEmitListener({ click: null }, 'onClick')).toBe(true)
-    expect(isEmitListener({ click: null }, 'onclick')).toBe(false)
-    expect(isEmitListener(['click'], 'onBlick')).toBe(false)
-    expect(isEmitListener({ click: null }, 'onBlick')).toBe(false)
+  test('merging from mixins', () => {
+    const mixin = {
+      emits: {
+        foo: (arg: number) => arg > 0
+      }
+    }
+    const Foo = defineComponent({
+      mixins: [mixin],
+      render() {},
+      created() {
+        this.$emit('foo', -1)
+      }
+    })
+    render(h(Foo), nodeOps.createElement('div'))
+    expect(`event validation failed for event "foo"`).toHaveBeenWarned()
+  })
+
+  test('.once', () => {
+    const Foo = defineComponent({
+      render() {},
+      emits: {
+        foo: null
+      },
+      created() {
+        this.$emit('foo')
+        this.$emit('foo')
+      }
+    })
+    const fn = jest.fn()
+    render(
+      h(Foo, {
+        onFooOnce: fn
+      }),
+      nodeOps.createElement('div')
+    )
+    expect(fn).toHaveBeenCalledTimes(1)
+  })
+
+  describe('isEmitListener', () => {
+    test('array option', () => {
+      const def1 = { emits: ['click'] }
+      expect(isEmitListener(def1, 'onClick')).toBe(true)
+      expect(isEmitListener(def1, 'onclick')).toBe(false)
+      expect(isEmitListener(def1, 'onBlick')).toBe(false)
+    })
+
+    test('object option', () => {
+      const def2 = { emits: { click: null } }
+      expect(isEmitListener(def2, 'onClick')).toBe(true)
+      expect(isEmitListener(def2, 'onclick')).toBe(false)
+      expect(isEmitListener(def2, 'onBlick')).toBe(false)
+    })
+
+    test('with mixins and extends', () => {
+      const mixin1 = { emits: ['foo'] }
+      const mixin2 = { emits: ['bar'] }
+      const extend = { emits: ['baz'] }
+      const def3 = {
+        mixins: [mixin1, mixin2],
+        extends: extend
+      }
+      expect(isEmitListener(def3, 'onFoo')).toBe(true)
+      expect(isEmitListener(def3, 'onBar')).toBe(true)
+      expect(isEmitListener(def3, 'onBaz')).toBe(true)
+      expect(isEmitListener(def3, 'onclick')).toBe(false)
+      expect(isEmitListener(def3, 'onBlick')).toBe(false)
+    })
+
+    test('.once listeners', () => {
+      const def2 = { emits: { click: null } }
+      expect(isEmitListener(def2, 'onClickOnce')).toBe(true)
+      expect(isEmitListener(def2, 'onclickOnce')).toBe(false)
+    })
   })
 })